added bill API branch for merge
