--- SharonsTest/views/index.handlebars
@@ -1,121 +1,0 @@
-<h1>Bills</h1>
-
-
-  {{#each billsData}}
-  <form style = "border: 2px solid"> 
-  <div class="form-group">
-    <label for="BillIDBillName"><b>{{StateBillID}}</b> : {{BillName}}</label>
-  </div>
-  <div class="form-group">
-    <label for="ActionDate"> {{ActionDate}}</label>
-  </div>
-    <div class="form-group">
-    <label for="Create"> {{Create}}</label>
-  </div>
-    <div class="form-group">
-    <label for="BillProgress"> {{BillProgress}}</label>
-  </div>
-  <div class="form-group">
-     <label for="Sponsors"><b>Sponsors</b> : {{Sponsors}}</label>
-  </div>
-  <div class="form-group">
-    <label for="Summary"><b>Summary</b>:</label>
-    <textarea class="form-control" id="Summary" rows="3">{{Summary}}</textarea>
-  </div>
-  <div class="form-group">
-    <label for="OfficialDocument"><b>More Info</b>: <a href={{OfficialDocument}}>{{OfficialDocument}}</a></label>
-  </div>
-</form>
-<br>
-  {{/each}}
-
-
-<button id="status"><a href="/active">Click me to see active stuffs!!</a></button>
-{{#each billsDataActive}}
-  <form style = "border: 2px solid"> 
-  <div class="form-group">
-    <label for="BillIDBillName"><b>{{StateBillID}}</b> : {{BillName}}</label>
-  </div>
-  <div class="form-group">
-    <label for="ActionDate"> {{ActionDate}}</label>
-  </div>
-    <div class="form-group">
-    <label for="Create"> {{Create}}</label>
-  </div>
-    <div class="form-group">
-    <label for="BillProgress"> {{BillProgress}}</label>
-  </div>
-  <div class="form-group">
-     <label for="Sponsors"><b>Sponsors</b> : {{Sponsors}}</label>
-  </div>
-  <div class="form-group">
-    <label for="Summary"><b>Summary</b>:</label>
-    <textarea class="form-control" id="Summary" rows="3">{{Summary}}</textarea>
-  </div>
-  <div class="form-group">
-    <label for="OfficialDocument"><b>More Info</b>: <a href={{OfficialDocument}}>{{OfficialDocument}}</a></label>
-  </div>
-</form>
-<br>
-  {{/each}}
-
-<script type="text/javascript">
-
-
-</script>
-
-
-{{!-- <script type="text/javascript">
-
-
-
-
-
-
-
-
-
-
-
-$(".delmovie").on("click", function(event){
-  var id = $(this).data("button")
-  event.preventDefault()
-  $.ajax("api/movies/" + id,{
-     type: "DELETE"
-  }).then(function(){
- console.log(data)
-    location.reload()
-  })
-})
-
-
-$("#newMovieForm").on("submit", function(event){
-  event.preventDefault()
-  var newMovie = {
-    movie: $("#insertNewMovies").val().trim()
-  }
-  $.post("/api/movies", newMovie).then(function(err, data){
-    console.log(data)
-    location.reload()
-  })
-})
-
-$("#updateMovies").on("submit", function(event){
-  event.preventDefault()
-  var id = $("#select").val().trim()
-  var updatedMovie = {
-    movie: $("#updatedMovie").val().trim()
-  }
-  $.ajax("/api/movies/" + id,{
-type: "PUT",
-data: updatedMovie
-
-  }).then(function() {
-        console.log("updated id ", id);
-        // Reload the page to get the updated list
-        location.reload();
-      })
-})
-
-
-</script> --}}
