+++ routes/apiRoutes.js
@@ -1,5 +1,7 @@
+
 //testing route for ajax
 module.exports = (app) => {
+
   app.post("/api/signup", (req, res) => {
     //create new db entry here
     console.log(req.body);
