+++ views/index.handlebars
@@ -1,33 +1,62 @@
-<div id="home">
-  <div class="container">
-    <div class="row">
-      <div class="col-8 offset-2">
-        <h1>{{msg}}</h1>
-        <h2>This is our Home Page!</h2>
-        <form class="clearfix mb-4" action="POST">
-          <div class="form-group">
-            <label for="example-text">Example Text</label>
-            <input type="text" id="example-text" class="form-control" aria-describedby="example-text" placeholder="Enter Example Text">
-          </div>
-          <div class="form-group">
-            <label for="example-description">Example Description</label>
-            <textarea class="form-control" id="example-description" aria-describedby="example-description" placeholder="Enter Example Description"></textarea>
-          </div>
-          <button id="submit" class="btn btn-primary float-right">Submit</button>
-        </form>
-        <h3>Examples:</h3>
-        <ul id="example-list" class="list-group">
-          {{#each examples}}
-          <li data-id="{{this.id}}" class="list-group-item">
-            <a href="example/{{this.id}}">{{this.text}}</a>
-            <button class="btn btn-danger float-right delete">ｘ</button>
-          </li>
-          {{/each}}
-        </ul>
-      </div>
-    </div>
-  </div>
-</div>
+<h1>Bills</h1>
 
-<script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
-<script src="/js/index.js"></script>
+
+  {{#each billsData}}
+  <form style = "border: 2px solid"> 
+  <div class="form-group">
+    <label for="BillIDBillName"><b>{{StateBillID}}</b> : {{BillName}}</label>
+  </div>
+  <div class="form-group">
+    <label for="ActionDate"> {{ActionDate}}</label>
+  </div>
+    <div class="form-group">
+    <label for="Create"> {{Create}}</label>
+  </div>
+    <div class="form-group">
+    <label for="BillProgress"> {{BillProgress}}</label>
+  </div>
+  <div class="form-group">
+     <label for="Sponsors"><b>Sponsors</b> : {{Sponsors}}</label>
+  </div>
+  <div class="form-group">
+    <label for="Summary"><b>Summary</b>:</label>
+    <textarea class="form-control" id="Summary" rows="3">{{Summary}}</textarea>
+  </div>
+  <div class="form-group">
+    <label for="OfficialDocument"><b>More Info</b>: <a href={{OfficialDocument}}>{{OfficialDocument}}</a></label>
+  </div>
+</form>
+<br>
+  {{/each}}
+
+
+<button id="status"><a href="/active">Click me to see active stuffs!!</a></button>
+{{#each billsDataActive}}
+  <form style = "border: 2px solid"> 
+  <div class="form-group">
+    <label for="BillIDBillName"><b>{{StateBillID}}</b> : {{BillName}}</label>
+  </div>
+  <div class="form-group">
+    <label for="ActionDate"> {{ActionDate}}</label>
+  </div>
+    <div class="form-group">
+    <label for="Create"> {{Create}}</label>
+  </div>
+    <div class="form-group">
+    <label for="BillProgress"> {{BillProgress}}</label>
+  </div>
+  <div class="form-group">
+     <label for="Sponsors"><b>Sponsors</b> : {{Sponsors}}</label>
+  </div>
+  <div class="form-group">
+    <label for="Summary"><b>Summary</b>:</label>
+    <textarea class="form-control" id="Summary" rows="3">{{Summary}}</textarea>
+  </div>
+  <div class="form-group">
+    <label for="OfficialDocument"><b>More Info</b>: <a href={{OfficialDocument}}>{{OfficialDocument}}</a></label>
+  </div>
+</form>
+<br>
+  {{/each}}
+
+<script type="text/javascript"></script>
